feat(api): 標準化 Auth API 回應格式
- 統一登入、註冊、刷新 token 回應為 { data: {...} } 格式
- 修正 RefreshToken 模型類型轉換問題
- 更新 API 回應結構以符合 RESTful 標準

+++ app/Models/RefreshToken.php
@@ -43,7 +43,7 @@ class RefreshToken extends Model
         $plainToken = Str::random(64);
         $hashedToken = hash('sha256', $plainToken);
 
-        $expirationMinutes = config('sanctum.refresh_token_expiration', 43200);
+        $expirationMinutes = (int) config('sanctum.refresh_token_expiration', 43200);
 
         $refreshToken = self::create([
             'user_id' => $user->id,
